+++ Download-OpenVPN-Configuration-Example/DownloadVpnCfg.properties
@@ -22,7 +22,7 @@
 
 ################################################################################
 # --Required--
-# HOST:                   WASaaS Broker API Host     (Example: https://wasaas-broker.ng.bluemix.net/wasaas-broker/api/v1)
+# HOST:                   WASaaS Broker API Host     (Example: https://wasaas-broker.us-south.websphereappsvr.cloud.ibm.com/wasaas-broker/api/v1)
 # ORGANIZATION:           Personal Organization name (Example: johndoe@ibm.com)
 # SPACE:                  Personal Space name        (Example: dev)
 # SERVICE_INSTANCE_ID:    The Service Instance ID containing the nodes (Example: dc8djk2-ddbf-43n33-ba4e-132094dn3imd)
@@ -33,8 +33,8 @@ HOST="YOUR_HOST_URL"
 ORGANIZATION="YOUR_ORGANIZATION"
 SPACE="YOUR_SPACE"
 SERVICE_INSTANCE_ID="YOUR_SERVICE_INSTANCE_ID"
-# either set BASIC_AUTH or 
-#BASIC_AUTH="Basic YOUR_BASIC_AUTHENTICATION_TOKEN"
+# either set BASIC_AUTH
+# BASIC_AUTH="Basic YOUR_BASIC_AUTHENTICATION_TOKEN"
 # or
-# use single use passcode on command line for federated id authentication
+# use a single use passcode on command line for federated id authentication
