+++ Download-OpenVPN-Configuration-Example/README.md
@@ -1,20 +1,20 @@
 # Download OpenVPN configuration Example Script
 
-This script allows consumers to download a configuration document for OpenVPN client access. 
+This script allows consumers to download a configuration document for OpenVPN client access.
 
 ## Running the Script
 
 * Populate Required Fields
   * The following fields must be populated in order to run the script
-      1. HOST         (Example: https://wasaas-broker.ng.bluemix.net/wasaas-broker/api/v1)                 
-      2. ORGANIZATION (Example: johndoe@ibm.com)          
-      3. SPACE        (Example: dev)           
+      1. HOST         (Example: https://wasaas-broker.us-south.websphereappsvr.cloud.ibm.com/wasaas-broker/api/v1)
+      2. ORGANIZATION (Example: johndoe@ibm.com)
+      3. SPACE        (Example: dev)
       4. SERVICE_INSTANCE_ID   (Example: dc8djk2-ddbf-43n33-ba4e-132094dn3imd)
 
   * Authentication. Either set BASIC_AUTH
       (5). BASIC_AUTH   (Example: Basic 99dj9jf9u7f77f7f7hwh3u4hjjnjk)         
       OR for federated ids use single use passcode on commandline.  Get one using browser from
-      E.g. https://login.ng.bluemix.net/UAALoginServerWAR/showpasscode.jsp
+      E.g. https://identity-1.us-south.iam.cloud.ibm.com/identity/passcode
 
 * Verify that execute permissions are enabled on the script
   * If execute permissions are not enabled run `chmod 755 ScaleService.sh`
