+++ README.md
@@ -1,15 +1,15 @@
-# WebSphere Application Server in Bluemix
+# WebSphere Application Server in IBM Cloud
 
-IBM [WebSphere Application Server in Bluemix][wasaas_docs_url] is a service in Bluemix that facilitates quick setup on a pre-configured WebSphere Application Server Liberty, Traditional Network Deployment, or Traditional WebSphere instance in a hosted cloud environment.
+IBM [WebSphere Application Server in IBM Cloud][wasaas_docs_url] is a service in IBM Cloud that facilitates quick setup on a pre-configured WebSphere Application Server Liberty, Traditional Network Deployment, or Traditional WebSphere instance in a hosted cloud environment.
 
-WebSphere Application Server in Bluemix provides consumers with pre-configured Traditional WebSphere and Liberty Profile servers. It is hosted on virtual machine guests with root access to the guest operating system. When you are creating your service, choose between Liberty, Traditional ND, or Traditional WebSphere.
+WebSphere Application Server in IBM Cloud provides consumers with pre-configured Traditional WebSphere and Liberty Profile servers. It is hosted on virtual machine guests with root access to the guest operating system. When you are creating your service, choose between Liberty, Traditional ND, or Traditional WebSphere.
 
 You can create and manage this service in 2 ways
 
-  1. In your browser by creating an instance via the [Bluemix Catalog][catalog_url]
-  1. Programmatically using the WebSphere Application Server in Bluemix *API*
+  1. In your browser by creating an instance via the [IBM Cloud Catalog][catalog_url]
+  1. Programmatically using the WebSphere Application Server in IBM Cloud *API*
 
-The example code in the WebSphere-In-Bluemix-API-Examples folder contains Java code that utilizes the API to perform tasks such as:
+The example code in the WebSphere-In-IBM-Cloud-API-Examples folder contains Java code that utilizes the API to perform tasks such as:
 
   * Create a service instance.
   * Read credentials for your machine.
@@ -23,5 +23,5 @@ The example code in the Scaling-Service-Example folder contains a script that ut
   * Start desired number of nodes within a service instance
   * Stop desired number of nodes within a service instance
 
-[wasaas_docs_url]: https://new-console.ng.bluemix.net/docs/services/ApplicationServeronCloud/index.html
-[catalog_url]: https://console.ng.bluemix.net/catalog/services/websphere-application-server/
+[wasaas_docs_url]: https://cloud.ibm.com/docs/services/ApplicationServeronCloud/index.html#about
+[catalog_url]: https://cloud.ibm.com/catalog/services/websphere-application-server
