+++ Scaling-Service-Example/ScaleService.properties
@@ -21,7 +21,7 @@
 
 
 ################################################################################
-# HOST:                   WASaaS Broker API Host     (Example: https://wasaas-broker.ng.bluemix.net/wasaas-broker/api/v1)
+# HOST:                   WASaaS Broker API Host     (Example: https://wasaas-broker.us-south.websphereappsvr.cloud.ibm.com/wasaas-broker/api/v1)
 # ORGANIZATION:           Personal Organization name (Example: johndoe@ibm.com)
 # SPACE:                  Personal Space name        (Example: dev)
 # BASIC_AUTH:             "Basic " + Base 64 Encoded (username + ":" + password) (Example: Basic 99dj9jf9uwh3u4hjjnjk)
