+++ WebSphere-In-IBM-Cloud-API-Examples/GetResources.java
@@ -5,20 +5,22 @@ import java.io.InputStreamReader;
 import java.net.HttpURLConnection;
 import java.net.URL;
 
-// Delete a single service instance.
-public class DeleteServiceInstance {
-	/* WebSphere Application Server for Bluemix API URL.
+// Get info about all resources associated with a service instance.
+public class GetResources {
+	/* WebSphere Application Server for IBM Cloud API URL.
 	 * Available Environments:
-	 * Dallas - https://wasaas-broker.ng.bluemix.net/wasaas-broker/api/v1
-	 * London - https://wasaas-broker.eu-gb.bluemix.net/wasaas-broker/api/v1
-	 * Sydney - https://wasaas-broker.au-syd.bluemix.net/wasaas-broker/api/v1
+	 * Dallas - https://wasaas-broker.us-south.websphereappsvr.cloud.ibm.com/wasaas-broker/api/v1
+	 * London - https://wasaas-broker.eu-gb.websphereappsvr.cloud.ibm.com/wasaas-broker/api/v1
+	 * Sydney - https://wasaas-broker.au-syd.websphereappsvr.cloud.ibm.com/wasaas-broker/api/v1
+	 * Frankfurt - https://wasaas-broker.eu-de.websphereappsvr.cloud.ibm.com/wasaas-broker/api/v1
 	 */
-	private static final String apiEndpoint = "https://wasaas-broker.ng.bluemix.net/wasaas-broker/api/v1";
+	
+	private static final String apiEndpoint = "https://wasaas-broker.us-south.websphereappsvr.cloud.ibm.com/wasaas-broker/api/v1";
 
 	public static void main(String[] args) throws IOException{
 		// You can see how to get your access token from GetOAuthToken sample class.
 		String accessToken = "<YOUR_ACCESS_TOKEN>";
-		// The Bluemix organization & space to query - case sensitive.
+		// The IBM Cloud organization & space to query - case sensitive.
 		String org = "<YOUR_ORG>"; // Example: johndoe@ibm.com
 		String space = "<YOUR_SPACE>"; // Example: dev
 		// You can see how to get a service instance ID from the GetServiceInstances sample class.
@@ -28,27 +30,29 @@ public class DeleteServiceInstance {
 		System.setProperty("https.protocols", "TLSv1.2");
 
 		// Create the URL.
-		URL orgsURL = new URL(apiEndpoint + "/organizations/" + org + "/spaces/" + space + "/serviceinstances/" + serviceInstanceID);
+		URL orgsURL = new URL(apiEndpoint + "/organizations/" + org + "/spaces/" + space + "/serviceinstances/" + serviceInstanceID + "/resources");
 		HttpURLConnection con = (HttpURLConnection) orgsURL.openConnection();
-		con.setRequestMethod("DELETE");
+		con.setRequestMethod("GET");
 		con.setRequestProperty("Authorization", "Bearer " + accessToken);
 
-		if (HttpURLConnection.HTTP_NO_CONTENT == con.getResponseCode()) {
-			System.out.println("Successfully Deleted.");
+		BufferedReader br = null;
+		if (HttpURLConnection.HTTP_OK == con.getResponseCode()) {
+			br = new BufferedReader(new InputStreamReader(con.getInputStream()));
 		}
 		else {
-			BufferedReader br = new BufferedReader(new InputStreamReader(con.getErrorStream()));
-			StringBuffer response = new StringBuffer();
-			String line;
-
-			while ((line = br.readLine()) != null) {
-				response.append(line);
-			}
-			br.close();
-
-			// Response from the request.
-			System.out.println(response.toString());
+			br = new BufferedReader(new InputStreamReader(con.getErrorStream()));
 		}
+
+		StringBuffer response = new StringBuffer();
+		String line;
+
+		while ((line = br.readLine()) != null) {
+			response.append(line);
+		}
+		br.close();
+
+		// Response from the request.
+		System.out.println(response.toString());
 	}
 
 }
